Split quick search button and modal
